Add a solution for day 11, part1
I don't know how to do the part2: that is compute the length after 75
blinks (instead of 25 as for part1).
The length grows exponentially, it's impossible to brute force this case.
May be that the length expression is a recurrent sequence of order
greater than 3. But there is nothing less certain.